refactor(Controller): 클래스명 수정 및 DTO 로 데이터 통신
