+++ src/main/java/chess/controller/ChessControllerAdvice.java
@@ -1,0 +1,23 @@
+package chess.controller;
+
+import chess.exception.ChessException;
+import org.springframework.dao.DataAccessException;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.ControllerAdvice;
+import org.springframework.web.bind.annotation.ExceptionHandler;
+
+@ControllerAdvice
+public class ChessControllerAdvice {
+    @ExceptionHandler(ChessException.class)
+    public ResponseEntity<String> handle(ChessException e) {
+        return ResponseEntity.status(e.code())
+                             .body(e.desc());
+    }
+
+    @ExceptionHandler(DataAccessException.class)
+    public ResponseEntity<String> handle(DataAccessException e) {
+        return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR)
+                             .body(e.getMessage());
+    }
+}
