+++ src/main/java/chess/service/ChessService.java
@@ -18,10 +18,8 @@ public class ChessService {
         this.commandDao = commandDao;
     }
 
-    public void move(Long gameId, MoveDto moveDto) {
+    public void move(Long gameId, String from, String to) {
         Game game = loadGame(gameId);
-        String from = moveDto.getFrom();
-        String to = moveDto.getTo();
         game.move(Position.from(from), Position.from(to));
         commandDao.insert(gameId, from, to);
     }
